renamed kernel_tuning_utils to reflect that this is for 3D conv
